Seperate Fun Fact resolver
