+++ client/src/components/FunFact.jsx
@@ -21,7 +21,6 @@ export default function FunFact() {
   useEffect(() => {
     const interval = setInterval(() => {
       // console.log("timer");
-
       //setInterval will be called when there is array of facts.
       if (data) {
         setIndex(Math.floor(Math.random() * data.funFact.length));
@@ -31,16 +30,17 @@ export default function FunFact() {
     // Return in useEffect to clear the interval when detached from reactDOM
     return () => clearInterval(interval);
   }, [loading, error, data]);
-  if (loading) return <p>Loading...</p>;
   if (error) return <p>Error : {error.message}</p>;
   return (
     <div>
       <Alert variant="success">
         <BookOpenCheck className="h-4 w-4" />
-        <AlertTitle>Did you know?</AlertTitle>
-        <AlertDescription className="text-sm">
-          {/*Displays fact of random index */}
-          {data.funFact[index]}
+        {!loading && <AlertTitle>Did you know?</AlertTitle>}
+        <AlertDescription className="text-md">
+          {/*Displays loading when loading else displays fact of random index */}
+          {loading
+            ? "Hold onto your hats... Here comes a wild tidbit for you!"
+            : data.funFact[index]}
         </AlertDescription>
       </Alert>
     </div>
